feat(calendar): binding calendar api
